Add tests for coverage
