+++ ui/package.json
@@ -46,6 +46,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "jest-fetch-mock": "^3.0.3",
     "react-test-renderer": "^18.3.1",
     "ts-jest": "^29.1.4"
   },
