+++ ui/src/App.tsx
@@ -6,31 +6,10 @@ import Sidebar from './components/Sidebar';
 import './App.css';
 import Home from './components/Home';
 import TextInputView from './components/TextInputView';
-import { FormData } from './components/FormComponent';
-import { BeamClient, TextPromptRequest } from "./api/beam";
 import { config } from '../../conf/sdConfig';
+import { beamApiCall } from './api';
 
-const beamAuthToken = config.beam.authToken;
 const beamModels = config.beam.models;
-
-const beamApiCall = (formData: FormData): Promise<string> => {
-  const request: TextPromptRequest = {
-      prompt: formData.prompt,
-  }
-  if (formData.height && formData.height.trim().length > 0) {
-    request.height = parseInt(formData.height);
-  }
-  if (formData.width && formData.width.trim().length > 0) {
-    request.width = parseInt(formData.width);
-  }
-  if (formData.negativePrompt && formData.negativePrompt.trim().length > 0) {
-    request.negative_prompt = formData.negativePrompt;
-  }
-  const beamAppId: string  = formData.model;
-  const beamClient = new BeamClient({ authToken: beamAuthToken });
-  return beamClient.textPrompt(beamAppId, request)
-    .then(taskId => beamClient.pollTaskStatus(taskId));
-}
 
 const Img2Img = () => <h2>Image to Image (Img2Img)</h2>;
