+++ ui/src/api/index.ts
@@ -1,1 +1,2 @@
 export * from './beam';
+export * from './apiCalls';
